#27 Put data classes in their own class file
