+++ Assets/Scripts/Corporation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 09a3da00042e46142a5a4f2b9ed4ccac
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
