+++ Assets/Scripts/Stock.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 35d2ee9aebac6a34fb03047b341016d7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
