+++ Assets/Scripts/Wallet.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 03d356ac8dacf594e81c81f5383f6b89
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
